Remove maxConcurrentOps, writesScheduler and some prefetches.
Remove maxConcurrentOps as it makes load tuning more difficult.
Instead, executor.maxInFlight could be used to limit the number of
concurrent queries.

Remove writesScheduler it is superfluous as writeReactive runs in it's
own executor already.  Observed that not much work was done in these
worker threads and the additional scheduling was slowing things down.

Remove opsBufferSize and other prefetches.  The only prefetch parameter
that seems to be important is the one passed to parallel().  The main
benefit of making this value larger is that the queue used to enqueue
tasks is made larger, which prevents contention.