Improve DefaultReadResultMapper logic
This commit reverts the logic in DefaultReadResultMapper to iterate over
the row's columns and find the corresponding field for each column,
rather than the opposite (iterate over fields, find the corresponding
column). This doesn't change much the overall algorithm but is more
straight-forward and makes DefaultReadResultMapper's logic the exact
opposite of DefaultRecordMapper's.