BULK-22, #398: Address code review comments
Specifically, in addition to a few minor clean-up tasks, I switched from
specifying the S3 credentials in the config to specifying them as query
params on the individual URLs. To alleviate the performance concerns
about creating a new client for each URL in a urlfile, I have added a
cache with a configurable size.

I also added warnings when someone passes S3 credentials in an insecure
way or when they use a urlfile with over 100,000 lines (which can cause
out-of-memory errors). This limit is pretty arbitrary, and may be
adjusted in the future.