BULK-22, #398: Support reading s3 urls (#399)
* BULK-22, #398: Bump minor version in preparation for new feature

* BULK-22, #398: Add support for reading from AWS S3 URLs

Motivation:

In order to efficiently upload a large number of records from AWS S3 to
Cassandra, dsbulk needs to be able to read from s3:// URLs.

Modifications:

The main change here is adding the S3URLStreamHandler(Provider) classes.
This makes reading from S3 URLs transparent to the rest of the
application in the same way that reading from stdin/stdout is. A
modifcation to DataStaxBulkLoader was required to be able to pass the
config into these new classes.

Result:

Users may now pass s3:// URLs to load commands and dsbulk will retrieve
the content from AWS S3 and load it into Cassandra.

Future Work:

For this use case, only reading from S3 was required, but adding support
for writing to S3 should be a fairly simple modification to
S3URLStreamHandler.getOutputStream().

* BULK-22, #398: Add documentation for new read-from-S3 feature

I also added some new documentation to CONTRIBUTING.md for other new
contributors who might not know some of the details of the local build
process.

* BULK-22, #398: Remove insecure S3 credentials config options

* BULK-22, #398: Address code review comments

Specifically, in addition to a few minor clean-up tasks, I switched from
specifying the S3 credentials in the config to specifying them as query
params on the individual URLs. To alleviate the performance concerns
about creating a new client for each URL in a urlfile, I have added a
cache with a configurable size.

I also added warnings when someone passes S3 credentials in an insecure
way or when they use a urlfile with over 100,000 lines (which can cause
out-of-memory errors). This limit is pretty arbitrary, and may be
adjusted in the future.

* BULK-22, #398: Use Caffeine instead of LRUMap for caching

Also update URL-query-parsing logic and document the reason for choosing
20 as the cache size.

* BULK-22, #398: Handle equivalent S3 credential queries